pyproject.toml
[tool.poetry]
name = "rag-app"
version = "0.1.0"
description = "Your description here"
authors = ["Jason Liu <jason@jxnl.co>", "Ivan Leo <ivanleomk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.1.0"
pydantic = "^2.0.2"
typer = "^0.9.0"
lancedb = "^0.6.1"
tqdm = "^4.66.2"
rich = "^13.6.0"
python-frontmatter = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rag-app = 'rag_app.cli:app'
